ml_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.10.4
python-multipart==0.0.21
pillow==10.4.0
pytesseract==0.3.13
pdfplumber==0.11.4
pymupdf==1.24.10
opencv-python==4.10.0.84
transformers==4.46.3
torch==2.5.1
